refactor(Auth): re - organised repo
