+++ src/app/[locale]/(unauth)/auth/forgot-password/page.tsx
@@ -1,8 +1,9 @@
 import { getTranslations } from 'next-intl/server';
 import React from 'react';
 
-import { AuthLayout } from '@/layouts/AuthLayout';
-import { ForgotPasswordForm } from '@/layouts/ForgotPasswordForm';
+import { AuthLayout } from '../_components/AuthLayout';
+
+import { ForgotPasswordForm } from './_components/ForgotPasswordForm';
 
 export async function generateMetadata({
   params: { locale },
