+++ src/locales/en.json
@@ -56,7 +56,7 @@
     "site": "Site",
     "my_schedule": "My Schedule",
     "books": "Stories",
-    "profile": "Profile",
+    "hubers": "Hubers",
     "help": "Help",
     "contact": "Contact",
     "join_our_social_community": "JOIN OUR SOCIAL COMMUNITY",
