+++ src/locales/vi.json
@@ -66,6 +66,7 @@
     "site": "Trang web",
     "my_schedule": "Lịch trình của tôi",
     "books": "Câu chuyện",
+    "hubers": "Huber",
     "profile": "Hồ sơ",
     "help": "Trợ giúp",
     "contact": "Liên hệ",
